+++ apdt/ml/__init__.py
@@ -8,6 +8,7 @@ Check https://github.com/Zhiyuan-Wu/apdt for more information.
 '''
 from apdt.ml.general import DataSet, TFModel
 from apdt.ml.tf_model import wavenet_weight, WaveNet, mlp_weight, MLP, lstm_weight, LSTM
+from apdt.ml.tf_model import cnn1d_weight, cnn1d
 from apdt.ml.tf_model import Transformer_ar_decoder_weight, Transformer_ar_decoder
 from apdt.ml.tf_model import _WaveNet
 from apdt.ml.utils import stacked_window