+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### V0.1.8
 - ml.TFModel 增加 1-D CNN
 - 增加 ml.batch_norm 和 ml.layer_norm 方法
+- 重做 TFModel.Transformer
 - 为 MLP 增加 bn 支持
 - ml.DataSet 增加 shuffle_train_only 参数