+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Change Log
 
+#### V0.1.9
+
 #### V0.1.8
 - ml.TFModel 增加 1-D CNN
 - ml.TFModel 增加 2-D CNN
-- 增加 ml.batch_norm 和 ml.layer_norm 方法
+- ml.TFModel 增加 deepsets
 - 重做 TFModel.Transformer
+- 增加 ml.batch_norm 和 ml.layer_norm 方法
 - 为 MLP 增加 bn 支持
 - ml.DataSet 增加 shuffle_train_only 参数
 - TFModel 支持定义多个 loss